+++ client/web/src/enterprise/code-monitoring/components/actions/ActionEditor.test.tsx
@@ -17,11 +17,17 @@ describe('ActionEditor', () => {
         idName: 'email',
         actionEnabled: true,
         toggleActionEnabled: sinon.fake(),
+        includeResults: true,
+        toggleIncludeResults: sinon.fake(),
         canSubmit: true,
         onSubmit: sinon.fake(),
         onCancel: sinon.fake(),
         canDelete: true,
         onDelete: sinon.fake(),
+        testButtonText: 'Send test email',
+        onTest: sinon.fake(),
+        testAgainButtonText: 'Send again',
+        testState: undefined,
     }
 
     test('expand and collapse with cancel button', () => {