+++ client/web/src/enterprise/insights/hooks/use-get-api.ts
@@ -3,9 +3,7 @@ import { useEffect, useState } from 'react'
 import { gql, useApolloClient } from '@apollo/client'
 
 import { IsCodeInsightsLicensedResult } from '../../../graphql-operations'
-import { CodeInsightsBackend } from '../core/backend/code-insights-backend'
-import { CodeInsightsGqlBackend } from '../core/backend/gql-backend/code-insights-gql-backend'
-import { CodeInsightsGqlBackendLimited } from '../core/backend/lam-backend/code-insights-gql-backend-limited'
+import { CodeInsightsBackend, CodeInsightsGqlBackend, CodeInsightsGqlBackendLimited } from '../core'
 
 /**
  * Returns the full or limited version of the API based on