+++ enterprise/internal/insights/compression/worker.go
@@ -34,14 +34,15 @@ type CommitIndexer struct {
 	db                database.DB
 	limiter           *rate.Limiter
 	allReposIterator  func(ctx context.Context, each func(repoName string, id api.RepoID) error) error
-	getCommits        func(ctx context.Context, db database.DB, name api.RepoName, after time.Time, operation *observation.Operation) ([]*gitdomain.Commit, error)
+	getCommits        func(ctx context.Context, db database.DB, name api.RepoName, after time.Time, until *time.Time, operation *observation.Operation) ([]*gitdomain.Commit, error)
 	commitStore       CommitStore
 	maxHistoricalTime time.Time
 	background        context.Context
 	operations        *operations
+	clock             func() time.Time
 }
 
-func NewCommitIndexer(background context.Context, base database.DB, insights dbutil.DB, observationContext *observation.Context) *CommitIndexer {
+func NewCommitIndexer(background context.Context, base database.DB, insights dbutil.DB, clock func() time.Time, observationContext *observation.Context) *CommitIndexer {
 	//TODO(insights): add a setting for historical index length
 	startTime := time.Now().AddDate(-1, 0, 0)
 
@@ -74,12 +75,14 @@ func NewCommitIndexer(background context.Context, base database.DB, insights dbu
 		background:        background,
 		getCommits:        getCommits,
 		operations:        operations,
+		clock:             clock,
 	}
+
 	return &indexer
 }
 
-func NewCommitIndexerWorker(ctx context.Context, base database.DB, insights dbutil.DB, observationContext *observation.Context) goroutine.BackgroundRoutine {
-	indexer := NewCommitIndexer(ctx, base, insights, observationContext)
+func NewCommitIndexerWorker(ctx context.Context, base database.DB, insights dbutil.DB, clock func() time.Time, observationContext *observation.Context) goroutine.BackgroundRoutine {
+	indexer := NewCommitIndexer(ctx, base, insights, clock, observationContext)
 
 	return indexer.Handler(ctx, observationContext)
 }
@@ -106,24 +109,38 @@ func (i *CommitIndexer) indexAll(ctx context.Context) error {
 	return nil
 }
 
-// indexRepository attempts to index the commits given a repository name. This method will absorb any errors that
-// occur during execution and skip the index for this repository.
+// maxWindowsPerRepo Limits the number of windows of commits indexRepository can process per run
+const maxWindowsPerRepo = 25
+
+// indexRepository attempts to index the commits given a repository name one time window at a time.
+// This method will absorb any errors that occur during execution and skip any remaining windows.
 // If this repository already has some commits indexed, only commits made more recently than the previous index will be added.
 func (i *CommitIndexer) indexRepository(name string, id api.RepoID) error {
-	err := i.index(name, id)
-	if err != nil {
-		log15.Error(err.Error())
+	windowsProccssed := 0
+	additionalWindows := true
+	// It is important that the window size stays consistent during processing
+	// so that it can correctly determine the time the repository has been indexed though
+	windowDuration := conf.Get().InsightsCommitIndexerWindowDuration
+	for additionalWindows && windowsProccssed < maxWindowsPerRepo {
+		var err error
+		additionalWindows, err = i.indexNextWindow(name, id, windowDuration)
+		windowsProccssed++
+		if err != nil {
+			log15.Error(err.Error())
+			return nil
+		}
 	}
 	return nil
+
 }
 
-func (i *CommitIndexer) index(name string, id api.RepoID) (err error) {
+func (i *CommitIndexer) indexNextWindow(name string, id api.RepoID, windowDuration int) (moreWindows bool, err error) {
 	ctx, cancel := context.WithTimeout(i.background, time.Second*45)
 	defer cancel()
 
 	err = i.limiter.Wait(ctx)
 	if err != nil {
-		return nil
+		return false, err
 	}
 
 	logger := log15.Root().New("worker", "insights-commit-indexer")
@@ -133,50 +150,63 @@ func (i *CommitIndexer) index(name string, id api.RepoID) (err error) {
 
 	metadata, err := getMetadata(ctx, repoId, i.commitStore)
 	if err != nil {
-		return errors.Wrapf(err, "unable to fetch commit index metadata repo_id: %v", repoId)
+		return false, errors.Wrapf(err, "unable to fetch commit index metadata repo_id: %v", repoId)
 	}
 
 	if !metadata.Enabled {
 		logger.Debug("commit indexing disabled", "repo_id", repoId)
-		return nil
+		return false, nil
 	}
 
-	searchTime := max(i.maxHistoricalTime, metadata.LastIndexedAt)
+	searchStartTime := max(i.maxHistoricalTime, metadata.LastIndexedAt)
+	commitLogRequestTime := i.clock().UTC()
 
-	logger.Debug("fetching commits", "repo_id", repoId, "after", searchTime)
-	commits, err := i.getCommits(ctx, i.db, repoName, searchTime, i.operations.getCommits)
+	var searchEndTime *time.Time
+	if windowDuration > 0 {
+		endTime := searchStartTime.Add(time.Duration(24*windowDuration) * time.Hour)
+		searchEndTime = &endTime
+	}
+
+	logger.Debug("fetching commits", "repo_id", repoId, "after", searchStartTime, "until", searchEndTime)
+	commits, err := i.getCommits(ctx, i.db, repoName, searchStartTime, searchEndTime, i.operations.getCommits)
 	if err != nil {
-		return errors.Wrapf(err, "error fetching commits from gitserver repo_id: %v", repoId)
+		return false, errors.Wrapf(err, "error fetching commits from gitserver repo_id: %v", repoId)
 	}
 
 	i.operations.countCommits.WithLabelValues().Add(float64(len(commits)))
 
-	if len(commits) == 0 {
-		logger.Debug("commit index up to date", "repo_id", repoId)
+	// default to thinking indexing is done
+	indexedThrough := commitLogRequestTime // The time we issued the git log request
+	moreWindows = false
 
-		if _, err = i.commitStore.UpsertMetadataStamp(ctx, repoId); err != nil {
-			return err
-
+	// If we are looking at a window of time determine if reached the end
+	if searchEndTime != nil {
+		moreWindows = searchEndTime.Before(commitLogRequestTime)
+		if moreWindows {
+			indexedThrough = *searchEndTime
 		}
-
-		return nil
 	}
 
-	log15.Debug("indexing commits", "repo_id", repoId, "count", len(commits))
-	err = i.commitStore.InsertCommits(ctx, repoId, commits, fmt.Sprintf("|repoName:%s|repoId:%d", repoName, repoId))
+	log15.Debug("indexing commits", "repo_id", repoId, "count", len(commits), "indexedThrough", indexedThrough)
+	err = i.commitStore.InsertCommits(ctx, repoId, commits, indexedThrough, fmt.Sprintf("|repoName:%s|repoId:%d", repoName, repoId))
 	if err != nil {
-		return errors.Wrapf(err, "unable to update commit index repo_id: %v", repoId)
+		return false, errors.Wrapf(err, "unable to update commit index repo_id: %v", repoId)
 	}
 
-	return nil
+	return moreWindows, nil
 }
 
 // getCommits fetches the commits from the remote gitserver for a repository after a certain time.
-func getCommits(ctx context.Context, db database.DB, name api.RepoName, after time.Time, operation *observation.Operation) (_ []*gitdomain.Commit, err error) {
+func getCommits(ctx context.Context, db database.DB, name api.RepoName, after time.Time, until *time.Time, operation *observation.Operation) (_ []*gitdomain.Commit, err error) {
 	ctx, endObservation := operation.With(ctx, &err, observation.Args{})
 	defer endObservation(1, observation.Args{})
 
-	return git.Commits(ctx, db, name, git.CommitsOptions{N: 0, DateOrder: true, NoEnsureRevision: true, After: after.Format(time.RFC3339)}, authz.DefaultSubRepoPermsChecker)
+	before := ""
+	if until != nil {
+		before = until.Format(time.RFC3339)
+	}
+
+	return git.Commits(ctx, db, name, git.CommitsOptions{N: 0, DateOrder: true, NoEnsureRevision: true, After: after.Format(time.RFC3339), Before: before}, authz.DefaultSubRepoPermsChecker)
 }
 
 // getMetadata gets the index metadata for a repository. The metadata will be generated if it doesn't already exist, such as
@@ -184,7 +214,7 @@ func getCommits(ctx context.Context, db database.DB, name api.RepoName, after ti
 func getMetadata(ctx context.Context, id api.RepoID, store CommitStore) (CommitIndexMetadata, error) {
 	metadata, err := store.GetMetadata(ctx, id)
 	if errors.Is(err, sql.ErrNoRows) {
-		metadata, err = store.UpsertMetadataStamp(ctx, id)
+		metadata, err = store.UpsertMetadataStamp(ctx, id, time.Time{}.UTC())
 	}
 	if err != nil {
 		return CommitIndexMetadata{}, err